Move DEBUG code path from main SceneDelegate to a new DebuggingSceneDelegate subclass to separate debug and test-specific code form production
